+++ Modelo.Analise.Api/Repository/Interface/IVendaRepository.cs
@@ -5,7 +5,7 @@ namespace Modelo.Analise.Api.Repository.Interface
     public interface IVendaRepository
     {
         Task<List<venda>> GetVendas();
-        Task<int> ObterQuantidadeDeVendasComparadoMesAnterior();
+        Task<decimal> ObterQuantidadeDeVendasComparadoMesAnterior();
         Task<List<venda>> ObterTopCincoVendas();
     }
 }
